test(toggle): migrate states
