+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-checked-ios-ltr-Mobile-Firefox-linux.png
