+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-checked-ios-rtl-Mobile-Chrome-linux.png
