+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-checked-ios-rtl-Mobile-Firefox-linux.png
