+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-checked-md-ltr-Mobile-Chrome-linux.png
