+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-checked-md-rtl-Mobile-Safari-linux.png
