+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-disabled-ios-ltr-Mobile-Chrome-linux.png
