+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-disabled-md-ltr-Mobile-Safari-linux.png
