+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-unchecked-ios-ltr-Mobile-Chrome-linux.png
