+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-unchecked-ios-rtl-Mobile-Firefox-linux.png
