+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-unchecked-ios-rtl-Mobile-Safari-linux.png
