+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-unchecked-md-ltr-Mobile-Firefox-linux.png
