+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-unchecked-md-rtl-Mobile-Chrome-linux.png
