+++ core/src/components/toggle/test/states/toggle.e2e.ts-snapshots/toggle-unchecked-md-rtl-Mobile-Safari-linux.png
